requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
matplotlib>=3.8.0
graphviz>=0.20.1
pillow>=10.0.0
numpy>=1.26.0
python-dateutil>=2.8.2
jira>=3.5.2
requests>=2.31.0
